ISSUE=9529 flow stat cache thread dies
In some circumstances the flow stat cacher thread is dying.  Its
related to a null pointer exception related to getProxy returning null
when the proxy is not found.  I fixed this in several locations.  While
trouble shooting we also found a potential synchronization issue with
the FlowSpaceFirewall.java class this.switches object not being
synchornized