there was an out of sync issue with flow stats
found an out of sync issue with flow stats when you change the path of
an OESS circuit and then diff.  The issue was that flows were still
associated to flows that were removed.  If we catch that case we then
find the new mapping and start updating that flow!