fixed the fact that flowstats were not checking actions
So a little re-architecting of the flow stat cacher.  Instead of trying
all the different permuations of a flow to find the slice it belongs to
we just store all the flows that are pushed to the switch and the flow
the controller sent.  If we don’t have the flow then we just add that
flow to our cache.  Added a unit test to verify the proper behavior