Set total_len when stripping a VLAN from pkt-in
The ofp_packet_in total_len field is not adjusted by any library
functions when we strip a VLAN tag from a data packet.  In addition,
deserializing and reserializing a packet might cause it to lose some
padding bytes.  To account for this, just set the total length to
whatever the length of the serialized packet is.